Support reading and writing multidimensional columns in ECSV format
